Correctly formatted all the code

+++ userland/bin/rmdir/rmdir.c
@@ -27,8 +27,8 @@
  * SUCH DAMAGE.
  */
 
-#include <unistd.h>
 #include <err.h>
+#include <unistd.h>
 
 /*
  * rmdir - remove a directory
@@ -37,15 +37,9 @@
  * Just calls the rmdir() system call.
  */
 
-int
-main(int argc, char *argv[])
-{
-	if (argc!=2) {
-		errx(1, "Usage: rmdir DIRECTORY");
-	}
+int main(int argc, char *argv[]) {
+	if(argc != 2) { errx(1, "Usage: rmdir DIRECTORY"); }
 
-	if (rmdir(argv[1])) {
-		err(1, "%s", argv[1]);
-	}
+	if(rmdir(argv[1])) { err(1, "%s", argv[1]); }
 	return 0;
 }
